Reduce vc-gen modules by adding invalidCreateVerifyData to helpers.
